even more cleaning. injective client types and services moved to appropriate folders
