+++ backend/src/bork-protocol/clients/injective-client/test/test-market-analysis.ts
@@ -1,11 +1,11 @@
-import { marketAnalysisService } from '../services/market-analysis-service';
-import type { TimeResolution } from '../types/market-history';
+import type { TimeResolution } from '../../../types/injective/market-history';
 import type {
   DerivativesMetrics,
   LiquidityAnalysis,
   OrderBookAnalysis,
   TechnicalAnalysis,
-} from '../types/technical-analysis';
+} from '../../../types/injective/technical-analysis';
+import { marketAnalysisService } from '../services/market-analysis-service';
 
 const MARKET_CONFIG = {
   type: 'spot' as const,
