+++ backend/src/bork-protocol/clients/x-client/accounts.ts
@@ -1,13 +1,13 @@
+import { tweetQueries } from '@/extensions/src/db/queries';
+import { KaitoService } from '@/services/kaito/kaito-service';
+import type { TwitterService } from '@/services/twitter//twitter-service';
+import { TwitterConfigService } from '@/services/twitter/twitter-config-service';
 import { initializeTargetAccounts } from '@/utils/accounts/account-initialization';
 import { selectTargetAccounts } from '@/utils/accounts/account-selection';
 import { selectTweetsFromAccounts } from '@/utils/accounts/tweet-selection';
+import { initializeTopicWeights } from '@/utils/topics/topics';
 import { processTweets } from '@/utils/tweet-analysis/process-tweets';
 import { type IAgentRuntime, elizaLogger } from '@elizaos/core';
-import { tweetQueries } from 'src/bork-protocol/extensions/src/db/queries';
-import { KaitoService } from 'src/bork-protocol/services/kaito-service';
-import { TwitterConfigService } from 'src/bork-protocol/services/twitter-config-service';
-import type { TwitterService } from 'src/bork-protocol/services/twitter-service';
-import { initializeTopicWeights } from 'src/bork-protocol/utils/topics/topics';
 
 export class TwitterAccountsClient {
   private twitterConfigService: TwitterConfigService;
