+++ backend/src/bork-protocol/services/twitter/twitter-spam.service.ts
@@ -1,6 +1,6 @@
+import { getUserSpamData } from '@/helpers/spam-helper';
 import { elizaLogger } from '@elizaos/core';
 import type { Tweet } from 'agent-twitter-client';
-import { getUserSpamData } from '../helpers/spam-helper';
 
 export interface SpamData {
   spamScore: number;
