+++ backend/src/bork-protocol/types/injective/technical-analysis.ts
